FEA: Allow multiple buckets with different settings
BREAKING CHANGE: Settings values are now specified as options in a dict, enabling the static and media stores to have different options, and enabling definition of any number of extra stores (each having a unique key). Storage can be instantiated to use any of these stores; the default being the MEDIA store.