DEP: Remove py3.8 support to avoid timezone difficulties with <=3.8
BREAKING CHANGE: We removed support for python 3.8, because we take advantage of the ZoneInfo management for python datetimes
that python 3.9 introduces. That allows us to fix a bug where non-naive datetimes would break the publish_time serialisation.

Yes, we could have backported everything. If you need that, we'll happily do so for a small donation to octue