FEA: Allow wrapping of execute by subclasses
BREAKING-CHANGE: This makes a further change to the api of the execute() method following the prior release. Now, the raw result is returned by execute() instead of a result, status tuple. The execute() method now accepts **task_kwargs instead of a raw request body requiring deserialisation. Deserialization and error handling are now managed where they ought to be - in the view.